test: move legacy tests into separate files for proper module mocking
